Remove help on workflow step switch
When the user navigates away from the build step, the build step help
should be removed from the help view.

I added a do:clear:help event to trigger this and a HelpView.clear method
as a handler for this event.  The do:clear:help event is then fired from
the BuildView.onHide method.

I also updated the help template to include a default message to be shown
if there is no help content set.

Closes #557