Detect initial loading of builder with hash ID
Detect the case when the user loads a story in the builder where the
builder ID is a hash argument.  In this case, the builder isn't properly
bootstrapped with story and section data.  Force a hard reload so the views
are properly bootstrapped.

Still to do is to properly generate the refresh URL in IE.

Addresses #417