Reload builder when story ID is in a hash
In IE9 (and presumably other browsers that don't support the History
API) Backbone falls back to using hashes to route to different views in
the application.  This however, causes a problem when reloading pages
because, without the ID in the URL, Django doesn't know to include the
JSON to bootstrap the Backbone views.

In the router, we detect this case, and force a reload of the page in
the browser (with the hash removed) to properly bootstrap the views.

Addresses #417