Integration test for #417
Test that the browser reloads the page when visiting a page with a hashed
story ID.

As part of this, I added some additional steps for clicking buttons to the
terrain module. I also updated the step for creating users to check if
a user already exists, allowing successive tests involving the same test
user to be run without clearing the database.