Don't refresh the browser on initial save
The build handler of the router now differentiates between the case
where the user is trying to save the story for the first time and where
they're trying to load an existing story.  In the former case, it's ok
to keep going as the story has already been bootstrapped by the
template. In the latter case, the browser should be reloaded with a
non-hashed story ID in the URL so the Django view can correctly
bootstrap the Backbone view.

Addresses #417