Show select2 placeholders
The placeholders for select2-enabled inputs were being hidden or cut off
in the Tag workflow step of the story builder.  Upgrading Select2 to version
3.2 seemed to correct this in Firefox, but not in Chrome. It appears that
the problem in Chrome is that Select2's resizeSearch function tries to
set the input width based on the width of the this.selection element
(in this particular view, it's a UL element with the .select2-choice class).
When the Select2 plugin tries to access the width of this element, it
detects it as 0, and so the input element is set to a width of 10px.
Setting a min-width on the input element (.select2-search-field input)
forces the width to be wide enough to show the placeholder text.

This commit also includes a fix that overrides the default height of
the .select2-search-field input.  The default height is 15px, which cuts off
half of the placeholder text in Chrome.

Addresses #245