Remove redundant site name in title
base.html appends the site name to the TITLE attribute and
story_detail.html was also prepending it. It should only be there once,
so I removed it from the story_detail.html template. Other than being
weird, one of the implications of the dual title was that it was making
it harder to fit tweets within the 140 char limits.

Addresses #421