New mirror for downloading Solr in CI build
The mirror used for downloading a more recent version of Solr as part of
the Travis CI build didn't work anymore.  Update to a working mirror.

Also, remove redundant command to install the solr-jetty package.  It was
being run in both a fabric task and in the ci_before_install.sh script.

Addresses #564