Resize figure to wrap iframe
Previously, the figure was only resized to fit images, but I noticed that
it looked a bit odd with video content as well. YouTube seems to be the
most common video provider, and it uses IFRAMEs, so I modified the
figure-resizing code to size the figure captions and borders to wrap
IFRAMEs as well.

Addresses #520.