Refactor model attribute normalization for views
We normalize the attributes of Story and Project models to be featured on
the homepage, so the same template (storybase/featured_object.html) can
be used for all items.  Previously, this normalization was handled in
the template tag functions (featured_stories, featured_projects), but
that caused a lot of repeated code.  For the updated homepage layout (#433),
Story, NewsItem and Project items will all appear in the same box, so
handling the normalization in a template tag becomes even uglier.  I decided
to move the normalization into a ``normalize_for_view`` method of the
model class.  This encapsulates the logic and underlying model attributes
from the view/template tag code.