Create a single featured items list
The homepage redesign (#433) calls for a single featured items list that
includes stories, news items, and projects.  Previously these were handled
by the ``featured_news``, ``featured_stories`` and ``featured_projects``
template tags.  These have been replaced by the ``featured_items`` tag
in the ``storybase_featured`` tag library.  In the future, if one wants to
break out a list of a particular type of featured items, the
``featured_items`` template tag takes an optional third argument which is
an iterable of objects to be displayed.

To show a list of featured stories, do something like this:

In the view code:

    context = {
        ...
        stories = Story.on_homepage.all(),
        ...
    }

In the template code:

    {% load storybase_featured %}

    <!-- Show 2 featured stories with a featured image widht of 334 -->
    {% featured_items 2 334 stories %}