Switch to higher contrast blue
Use #2a88db for the ``@highlight_blue`` blue in the stylesheets.

This commit also cleans up the explorer.less and widget.less files to use
the color variables when possible rather than hard-coded hex values.
Some of the colors in explorer.less were left as hex values because they
were closely tied to a color in an image and I didn't want to change edit
the images as well.

Addresses #449