Add admin user fixture
The story_templates.json fixture assumes that an admin user with pk=1 has
been created and uses this value for assigning the author field of the
template stories.  However, no such user exists when loading the fixture
in the unit tests.  Django 1.3.1 didn't seem to care about this, but
Django 1.4.3 displays an error message.  I created a fixture that creates
a single admin user to suppress the error.

Addresses #35