Change the story license immediately
In the publish/share view of the builder, the story license (and the
Creative Commons image and link) change as soon as the radio buttons are
changed, eliminating the need for a submit button.

In order to make this work, the Story needs to have a default license value.
The Story model has been updated so that all stories have a 'CC BY' license
when instantiated.

I also removed the padding on the forms in the Publish/Share view, made the
intermediate text (the text shown before the CC image and link are loaded)
indicate that more data is being loaded, and added a fallback message for
when the request to the CC API fails.

Closes #517.