Add numbers to the publish view steps
Add numbers to the publish view steps. I did this by making each
sub-view's element a ``li``.  I had to do a bit of CSS trickery to style
the ordered list numbers to match the headings.
http://www.456bereastreet.com/archive/201105/styling_ordered_list_numbers/
was very helpful with this.

I still need to test this under IE.

Addresses #515