Use LESS variables for publish step number styles
This is a tweak to the changes made in 2a8bcadaa92e1417c09d52e3053383b12115f325
that uses LESS variables to make the publish step numbers match the
header styles instead of hard-coding values.  This should make the code
more future proof.

I added a ``@styled-header-family`` variable to variables.less and also
use this variable to set the font in the ``.styled-header`` mixin in
base.less.

Addresses #515