Refactor publish step views
There was too much logic in the publish step templates and too much
logic related to toggling various elements based on publication status
in PublishView.  I broke most of the UI elements such as the publish
button, the to-do list and the view/publish buttons into their own views
and broke out common utility methods and initialization to a base view
class.

Addresses #515