Add ability to preview unpublished news items
Admins can now create news items with a status of "staged" and see them
on the home page. These will not be visible to non-admin users.

In order to do this, I created a new status of "staged" and modified
FeaturedManager.on_homepage to return these items as well as "published"
ones.

I then added a 'view' permission, implemented through user_can_view and
anonymoususer_can_view methods on the various model classes (actually
implemented in mixins).

Finally the featured_items template tag uses has_perm to check that a
user has the permission to view the item before passing the item to the
template.  I had to pass the context to this template tag in order to give
it access to the user.

Closes #593

Addresses #433, #567