Upgrade Backbone and Underscore
Upgrade Backbone and Underscore libraries to most recent version.

Backbone goes to version 0.9.10 and Underscore goes to version 1.4.3.

This is mostly because I wanted to start binding to events on models and
collections, when possible, instead of the event bus.  As of Backbone 0.9.9,
there are now listenTo and stopListening methods on Events, which make
binding event callbacks a little cleaner.

Note that in this version of Backbone, the Backbone object extends
Backbone.Events so it can be used as a global event bus.  Doing this might
make the code less verbose in the future.

I had to change the Asset model validation test to reflect the way that
validation doesn't automatically happen on Model.set and that invalid data
sends the "invalid" event rather than "error".