Use jQuery to detect element width
In ``ViewerApp.sizeFigCaption``, ``this.width`` was being used to detect
if the element had a width.  This was causing the caption and the bounding
element to be incorrectly sized (or even have the size set to 0) in
``ViewerApp.sizeFigCaptions``.  Checking the width with jQuery's ``width``
method seems to make things work correctly.  The ``width`` method returns
the current computed width of the element, which is what we want rather
than the element's raw width value.

Addresses #601