Properly populate preview bootstrap section list
When there were no published connected stories, the sections Backbone
collection wasn't getting bootstrapped with the draft connected stories
when in preview mode.  By passing an optional iterable of connected stories
to the ``BaseStructure.stories_json`` method, the draft connected stories
can trigger the inclusion of the connected stories list in the viewer
navigation when previewing a story. The logic for the getting the correct
connected stories iterable is done in ``StoryViewerView.get_context_data``.

Addresses #601