Trigger a 'set:featuredasset' signal
I want to move away from relying on the event bus (this.dispatcher in most
views) and try to trigger/listen-to events on the models and collections
themselves when that's appropriate.

To this end, the Story.setFeaturedAsset method now triggers a
'set:featuredasset' signal when the featured asset is set.