Implement switching between featured asset views.
Implement the views and logic that display the correct view depending on
whether the user is viewing, selecting or adding the featured image.

Also in this commit, I fixed a bug that I think is related to the
Backbone upgrade.  SectionEditView now only renders its subviews on a
"sync" signal, not on both "sync" and "reset".  I believe that in the
new Backbone versions, both events are triggered when a collection is
updated from the server, causing extra copies of the SectionAssetEdit
subviews to be created.