Don't show connected stories in "Latest Stories"
In order to implement this, I changed storybase.utils.latest_context to
take a Queryset as the first argument instead of a Model class. This lets
me do more filtering on the per-model side.

I also updated storybase.utils.latest_context to sort the lists by the
last_edited field instead of the published field.  This makes it consistent
with the sorting on the Explore page and on the Project and Organization
list views.

Addresses #609.