Hide connected story viewer and detail views
Detail and viewer views don't make sense for connected stories because
they're only meant to be shown in the story viewer of their "seed" story.
So, accessing their detail and viewer views should respond with a 404 error.

I also updated the "My Stories" view so that the title of connected stories
no longer links to the detail page, but shows an icon to help identify it
as connected and links to the seed story's detail page.