Forefront connected stories in latest stories list
Connected stories shouldn't be shown in the latest stories list, but they
should be forefronted somehow.  The way I've decided to do this is to
use a signal handler to update the last_edited field of a Story model when
a connected story is published. That way, stories with new connected stories
will be bumped to the top of the "Latest Stories" list on the homepage.

I also added a small link and icon reporting the number of connected
stories that links to the connected stories section of the detail page.

In order to do this, I added extra context to the the dictionary returned
by ``Story.normalize_for_view`` and updated the ``latest_stories`` template
tag to use the ``storybase_story/latest_stories.html`` template instead
of ``storybase/latest_objects.html``.  I added an ``extra`` block to the
``storybase/latest_objects.html`` template so templates that wanted to add
extra stuff could just inherit from ``storybase/latest_objects.html``.

Finally, I wired in a Google Analytics event to track how many time users
click this link to see if it's worthwile to take up space.

Closes #609