Remove checkboxes from legal view
The checkboxes aren't neccessary since the user has already agreed to them
when selecting the terms of service.  I've removed the checkboxes and, per
the suggestion by @jwirfs-brock, moved the view from outside the list of
required steps and into the top of the Publish/Share view.

As part of this change, I updated ``PublishViewBase`` to use the listener
initialization pattern used in the featured image selection views. Since
I override the listener initialization method, it seemed to make sense to
make it a public method, so I (re)named methods that perform this work
``initListeners`` throughout the codebase.

Addresses #590