Fix for #611
When a "seed" story was saved, its connected story relationships were being
clobbered.  This was because a non-connected story model was being
bootstrapped with an empty set of relations and this was always sent to
the server when the story was saved.

The fix for this involves properly bootstrapping the story relations in
the Django view and also only saving the relations to the server in the
case of new connected stories.

This fix saves a request in most cases.