Set the default featured image
Set the default featured asset from an image asset added to the story.
This mirrors the logic that occurs when retrieving the featured image on
the front end.

This happens via event listeners in ``FeaturedAssetView`` that checks when
the first image asset is added to the story and sets it as the featured
image.

To work around a race condition, I started having ``FileUploadMixin``
trigger a ``fileupload`` event on the model once it has successfully had
its related file uploaded.

Addresses #515.