Filter unwanted stories from banner
Use the recently created custom methods on ``StoryQuerySet`` and
``StoryManager`` to eliminate connected stories, template stories and
unpublished stories.

Since this filtering is pretty common, I added a ``public`` method to
``StoryQuerySet``/``StoryManager`` that performs this filtering and it's
this ``public`` method that I use in the various ``get_objects``
implementations.

Addresses #198