Use StoryQuerySet methods
When possible, use custom methods on StoryQuerySet to deal with
connected stories.

In the future, it might be possible to get rid of some of the methods
I put on story instances for dealing with featured stories now that the
query API is cleaner, but leave it in place for now.