Scale banner image cropping height on resize
Set the banner image container ('.images a') height on DOM ready and
window resize. Previously, only the width was set, but we need to set the
height as well to properly crop tall images.

I also added some comments to clarify the way that the CSS and jQuery event
handlers work together.

Addresses #198