Move "View All Stories" button
On project and organization detail pages, move the "View All Stories"
button higher up on the page, below the item summary/desription.

I also added a ``view-all-stories`` class to the button (actually it's
an ``a`` element with the ``button`` class) to make the element easier
to select. I then added a Google Analytics event to track clicks for
this.

Finally, I removed some unused jQuery plugins from the ``extrajs`` block
of the ``storybase_user/detail_base.html`` template.

Closes #531