Show link to connected stories in explorer
Show a link to the connected stories in the explorer.  In order to do
this, I added a read-only ``connected_count`` field to the StoryApi
response. This is generated by the ``Story.connected_count`` utility
method.

As the number of stories with connected stories is somewhat small, there
doesn't seem to be any noticeable performance penalty for the extra
database call to get the number of connected stories. If this changes,
we could cache the value and invalidate using a signal, similar to how
other ``ManyToMany`` fields on the story model are handled.

Closes #629