Supress error for many assets, same container
Catch the MultipleObjectsReturned execption, log an error, and pick one
of the multiple assets so the viewer sees something.

I had to add the section model instance to the context when calling the
``container`` template tag so I could include the section ID in the error
log and make it easier to look up and fix stories when this occurs.

Addresses #534