Section landing page template
This commit adds the ``cms_section_landing.html`` template, a template for
Django CMS pages that live at the top of their section.  By default, it displays
the teaser of the page and lists all the child pages along with their teasers.
These regions can be overridden by specifying plugins for the placeholders.

In order to access the teaser in the template, I had to patch the ``Page``
 model with a ``get_teaser`` method and patch the ``PageAttribute``
class (which implements the ``page_attribute`` template tag so it considers
``teaser`` and allowed attribute.

Finally, I made a custom manager for ``Teaser`` objects to implement the same
language fallback pattern as the native ``Title`` objects.

Addresses #426