A more dynamic, DRY megamenu template
These commits try to clean up the megamenu templates to avoid repeated
markup and hard-coded URLs.

I made the ``menu/megamenu.html`` template be included using the
``show_menu`` template tag instead of ``include``.  This allows the URLs
for the menu items to be generated from the back-end instead of
hard-coded in the templates.  This also lets us rely on the menu system
to tell us if the menu item is active instead of matching megamenu and
body classes (though I left in the wiring for the body classes).

I added a ``STORYBASE_MEGAMENU_ITEMS`` setting to specify which items
are in the megamenu and a template filter, ``megamenuitems``, to reduce
the ``children`` list passed to the template by ``show_menu`` to just
the desired items.

I broke each megamenu item's additional content into separate templates
that all inherit from ``menu/megamenu_extra_base.html``. See the comment at the
top of the ``menu/megamenu.html`` template for how this is wired up. By
default, these templates show the sub-pages and the top-level page's
teaser attribute.

I made the ``CMS_MENU_TITLE_OVERWRITE`` ``True`` by default in the
``settings`` module, allowing admins to specify a different title for
pages when they're displayed in the menu.

Closes #664