Render top and footer navigation from the menu
Instead of hard coding link text and URLs in the upper-right-hand corner
and footer, menus, generate them using the ``show_menu`` template tag.

I like this because it breaks out chunks of markup in ``base.html`` a
bit more and should get rid of annoying execptions when a page doesn't
exist in the CMS.

It also makes the footer menu automatically match the megamenu.