Clean up SectionAssetEditViews
``SectionAssetView.renderAssetViews`` created new SectionAssetEditView
instances each time it was called.  That seemed not only wasteful, but also
left ghost views that still responded to events.  This commit keeps track
of the child views and only creates new views when needed. It also calls
the ``close`` method on the sub-views to clean up their event bindings.

The sub-views for asset editing are tracked through a ``_assetEditViews``
object, with container names for keys.  The creation of
``SectionAssetEditView`` instances was also factored out of
``SectionAssetView.renderAssetViews``

Updated ``SectionEditView.removeAsset`` to remove the
``SectionAssetEditView`` when an asset is removed from the section.
``removeAsset`` gets called by ``changeLayout``, so the view is cleaned up
when the layout is changed as well.

I also had to update ``changeLayout`` to handle the case where there are no
asset models bound to the ``SectionAssetEditView`` instances to just remove
all the sub-views.

Finally, I renamed ``SectionAssetEditView.remove`` to
``handleRemove`` because it was clobering Backbone's ``View.remove``.

Addresses #671