so i'm having some trouble getting this to do quite the right thing, even in the scenario wherein we don't try to recover  the story template's placeholder title/suggested title on load. summary of behavior:
- when creating a new story from a template, the first section title gets its template-defined placeholder
- when creating a new section, that section gets a generic 'Section Title' placeholder
- when loading a saved story in which a section lacks a title, that section isn't getting *any* placeholder at the moment.

that last could be handled by the handlebars template, but perhaps we'd prefer it to live in the view?

also i updated the section list view: when a section has no title it displays 'Untitled Section' rather than 'New Section' ... seemed a little more appropriate. however, this is also handled in the handlebars template rather than in the view code.