Deal with transaction issues in tests
Since we catch the ``IntegrityError`` that occurs when we violate the
``unique_together`` constraint (see #535) both in ``SectionAssetResource``
and in some of the tests, this causes problems with further database calls
because of how ``TestCase`` handles transactions.  In one case, we can
use ``TransactionTestCase``, but in other cases, we have to not do the
additional database calls because Tastypie's ``ResourceTestsCase`` inherits
from ``TestCase`` so we can't get the ``TransactionTestCase`` functionality.
Luckily, this didn't really affect the value of the test.

Addresses #535