Clean up fix for #487
Make the assignment of the ``context`` key in the view context more
pythonic.

The assignment of the context in the code to create the viewer Backbone
view can just be a local variable.  We don't need to include it in the
namespace object.

Update the Changelog to reflect these changes.