Model.sync using multipart/form-data
Adds ``FileUploadModel``, a subclass of ``Backbone.Model`` that can send
the model attributes to the server as multipart form data instead of JSON.
This allows for uploading files in a single POST. This is handled by
overriding ``Model.sync``.  Currently, this only supports browsers that
implement the ``FormData`` and ``FileReader`` interfaces.  Next up is
adding support for other browsers by sending the form data via a hidden
IFRAME.  Once that's done, we can remove the ``FileUploadMixin`` view
mixin.

Make the ``Asset`` model extend ``FileUploadModel`` to gain this
functionality.

Update ``SectionEditView`` and ``SectionAssetEditView`` to take a ``language``
option.

Update ``SectionAssetEditView.processForm`` and
 ``SectionAssetEditView.saveModel`` to use the enhanced ``Asset`` model
implementation.

Clean up the code a bit to remove ``console.debug`` calls and to use ``view``
instead of ``that`` when returning a view.

Addresses #595