Remove FileUploadMixin
The ``FileUploadMixin`` view mixin is no longer needed now that
files can be uploaded via the multipart form data sync implemented
in ``FileUploadModel``. This commit updates the data set creation
view and the featured asset upload view to rely on the new sync
implementation.

On the backend DataSetResource had to be updated to be able to
accept the multipart file upload.

Found and fixed a bug with how the form was being reinitialized in
``FeaturedAssetAddView``. The old form is now removed and replaced
by a new one.

Now that ``AssetResource`` and ``DataSetResource`` can accept file
uploads, there's no need for the separate upload endpoints.

Addresses #595