Add weight field to Story model
Add a ``weight`` field to the ``Story`` model to allow for sorting by
a generated value.  This is implemented by making ``Story`` inherit from a
new abstract model class, ``WeightedModel``.

``WeightedModel`` provides a ``get_weight`` method that is overridden to
return the current weight value.  In the case of ``Story``, this is the
``published`` date/timestamp as an epoch.

The weight field allows for sorting stories by its published date/time while
taking into account the published date/time of connected stories. This
implementation uses signals to set the weight field to either the date the
story was published, or the date of the most recently published connected
story.

Addresses #625