Sort lates * on homepage by published date
Sort the latest stories, projects and organization lists on the homepage
by published date/time instead of last updated date/time.

This gets around the problem of the latest lists being changed when an
item's ``on_homepage`` flag is changed.  For stories, the ``weight``
field is used instead of ``published`` so the sorting can reflect connected
stories as well.

Addresses #625