Correctly set height and width of section list
Set the height and width of the section list factoring in the spacers. I
think that this wasn't adjusted when the design was updated to always show
the spacers.  So, when there are enough sections, the width of the element
won't be wide enough to contain all the sections and the element's height
grows.

Closes #656