Fix failing test
So this wasn't a race condition, just some silliness on my part. The story
saves were happening in within less than one second of each other, so when
they were converted to an epoch by get_weight and converted to an integer
for storing in the weight field (an IntegerField), the weight values are
the same. If the comparisons were <= instead of <, the tests would pass,
but since the comparisons used <, they were failing.

I don't think having duplicate weight values for closely published stories
is a problem at all, but I feel like this test should reflect the case
where stories are written in a specific order and have distinctly different
weights.  I updated the call to sleep to ensure that the saves occur at
least one second apart.

Addresses #625