Pass JavaScript error details using extra
Instead of serializing the additional JavaScript error data into the error
message, pass it to the logger using the ``extra`` keyword argument.  This
makes the error message clearer but still allows the additional information
(URL of script or view where error occured, line number in script or page,
username, browser location where script that triggered error is included)
to be visible to certain handlers.  In particular, the Sentry handler
displays the extra parameters (and other info) quite nicely in the
Sentry panel.