s/block-inline/inline-block/
The position of the workflow step tabs in the builder was off for the
connected builder because the position of one of the elements was
incorrectly specified as "block-inline" instead of "inline-block".

I also adjusted the positioning of the tabs so that the line separating the
items (created using a border property) was not at the very top of the
black bar.  I achieved this by using a margin-top property on each li
element instead of padding-top.

Closes #695