s/rollback/rollback_unless_managed/
When creating a new SectionAsset model instance fails because of an
IntegrityError because there's already an asset assigned to the specified
section and container, the transaction needs to be rolled back.  However,
this is only needed for testing because normally autocommit mode is used.
Using django.db.transaction.rollback_unless_managed instead of
django.db.transaction.rollback seems to do the trick and only tries to
do a rollback if autocommit isn't being used.

Addresses #595