Refresh asset views after conflicting asset error
When the user tries to specify asset content for a section/container that
already has content assigned, likely because the builder is open in
multiple tabs/windows, refresh the section's assets collection and the
asset views to show the asset specified in the other window (see #535).

Also fix a bug (#696) where the asset views don't appear after switching
to a different workflow step and switching back to the build step.