Set cookie to supress tour on popup close
Previously, the cookie set to supress showing the tour is only set when the
user clicks the close button on the final guider window. It is now set
when the user clicks the final close button, or the "X" icon in the
upper-right-hand corner of any of the tour popups.

This was annoying me when doing integration testing.  Furthermore, I believe
that users don't want to see the tour again if they close out of it early.
The tour remains accesible through the tour icon in the help drawer.

This commit also includes a cleanup of the guider construction by creating
an object with default options for guider construction. This is applied to
the arguments to ``createGuider`` by using ``_.defaults``.

Closes #697