Add close icon to alerts
Add a close icon that will remove an AlertView's element from the DOM.

Move the fade out functionality inside AlertView.

Add an onClose callback to the options of AlertView.

Set a cookie when the unsupported browser message is closed so the message
is not shown again when the user has closed it.

Addresses #548