Better throttling of alert messages
Instead of only tracking the last alert messages, keep track of all visible
messages and only show a message if it isn't already visible.  This prevents
the stack of alerts that occurs when the layout is changed.

Also add a ``sticky`` option to the ``showAlert`` method that keeps an alert
message at the top of the list. This improves the behavior of the unsupported
browser message.

Addresses #548