Fallback to mock geocoder
Yahoo has discontinued its Placefinder API, so we're switching to a
different geocoder that requires an API key. I can't think of a good way
to pass the API key to Travis CI without storing it in the repo. So,
I implemented a mock geocoder class that will be used if no geocoder is
specified via the ``STORYBASE_GEOCODER`` setting.

It's probably a good idea to mock the geocoder anyway so tests don't fail
if the endpoint goes down. In the case of the mock geocoder, we're just
testing that the routing to the geopy-based geocoder works correctly and
not the precision of the geocoder.

This fallback method allows us to run the tests with a specific geocoder
locally.

Addresses #700