Use icons to indicate saving state
As suggested by @patternleaf, instead of turning the button green to
indicate a successful save, use a check mark icon.  This avoids the issue
raised by @jwirfs-brock about the green color compelling the user to click
the button when it's in the "saved" state.

This commit also introduces a "saving" state which the button goes into
when clicked to give the user some kind of instant feedback. The button
does not go into the "saving" state during an autosave because there isn't
presently any signal sent for all of the different content types to
indicate that a save is being initiated.

Addresses #548