Apply site-wide form styles to proximity search
In the explore view, the proximity search text input and buttons should
look like those elsewhere on the site.  This is achieved by changing the
container element of the inputs from a ``div`` to a ``form``, triggering
the styles in ``forms.less``.

I also updated the selectors for button styles in ``forms.less`` so they
would match inputs of type button and reset in addition to submit.

Finally, I cleaned up the selectors of the styles for the proximity search
form in ``explorer.less`` by using LESS' nested rules.

Addresses #706