Factor template handling into base view
Factor template handling in Backbone views into a base view,
``storybase.views.HandlebarsTemplateView``.  This is implemented through
``storybase.views.HandlebarsTemplateMixin`` which allows the view handling
capabilities to be mixed in to objects other than views.

This approach makes template handling more consistent and makes it easier
to swap out template engines if we ever decide to abandon Handlebars.

``HandlebarsTemplateView`` takes a ``templateSource`` property which is
either a string containing the template markup, an object with keys
mapping to template strings, or a function returning a template string or
hash of template strings.  When specifying an object, the main template
is identified with a key of ``__main``.

A compiled version of the main template is assigned to ``this.template``.
If an object was provided for ``templateSource``, a particular compiled
template can be retrieved using ``this.getTemplate(key)``.

Addresses #707