Avoid race condition when updating index
Signal handlers are used to update the search index when SectionAsset
instances are deleted.  This works fine when they are deleted
explicitly, but causes a problem when they are deleted via the cascade
when the related Section is deleted.  When the section is deleted, the
story can't be fetched via ``instance.section.story``.  To work around
this, story a reference to the story in a ``pre_delete`` signal handler
so its available in the ``post_delete`` signal handler.

Addresses #723