Disable save button on 'do:save:story'
This fixes a race condition where the auto-save when leaving
the title text input and the save initiated by clicking the "Save"
button both initiate the initial save of the story via a POST to the
API.  This causes 2 story records to be created.

This commit updates ``LastSavedView`` so the state of the button is
updated on the ``do:story:save`` event rather than the click event.

So, when the autosave triggers the ``do:story:save`` event, the
button's state will be updated and disabled so it will not respond
to the click event.

Addresses #728