Update tastypie dependency to 0.9.15
Prior to this version, there was a weird implementation of
``RelatedField.should_full_hydrate`` that made it difficult to
use resources with related fields outside of a direct request.
In our case, this happens when we instantiate resource classes to
generate JSON to bootstrap the Backbone views in
``StoryBuilderView``.

I had previously implemented a custom field class that patched the
wonky ``should_full_hydrate`` method, but it seems best to just
require working version of Tastypie

Details about the problem and a reference to the commit that fixed
it are in the issue at
https://github.com/toastdriven/django-tastypie/issues/917